backend/pyproject.toml
[project]
name = "helloclaw-backend"
version = "0.1.0"
description = "HelloClaw Backend - AI Agent powered by HelloAgents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "hello-agents==1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "sse-starlette>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["api", "services", "tools", "channels", "workspace"]
